LPS-86974 FragmentEntryLink content moved to new component with al its logic
