+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/components/fragment_entry_link/FragmentEntryLinkContent.soy
@@ -1,0 +1,16 @@
+{namespace com.liferay.layout.admin.web.FragmentEntryLinkContent}
+
+/**
+ * FragmentEntryLink
+ */
+{template .render}
+	{@param? fragmentEntryLinkId: string}
+
+	<div
+		class="drag-fragment fragment-entry-link-content" ref="content"
+		data-fragment-entry-link-id="{$fragmentEntryLinkId}"
+	>
+		<div class="loading-animation"></div>
+	</div>
+
+{/template}
